Restructured agave lib to allow for compound methods in addition to the standard api
