+++ handlers/jobs.js
@@ -46,11 +46,11 @@ let handlers = {
      * List Apps
      */
     listApps(req, res, next) {
-        agave.listApps((err, resp) => {
+        agave.api.listApps((err, resp) => {
             if (err) {return next(err);}
             let apps = [];
             async.each(resp.body.result, (app, cb) => {
-                agave.getApp(app.id, (err, resp2) => {
+                agave.api.getApp(app.id, (err, resp2) => {
                     if (resp2.body && resp2.body.result) {
                         apps.push(resp2.body.result);
                     }
@@ -90,7 +90,7 @@ let handlers = {
                         return next(error);
                     }
 
-                    submitJob(job, (err, resp) => {
+                    agave.submitJob(job, (err, resp) => {
                         if (err) {return next(err);}
                         res.send(resp);
                     });
@@ -125,7 +125,7 @@ let handlers = {
             }
 
             // re-submit job with old job data
-            submitJob(job, (err, resp) => {
+            agave.submitJob(job, (err, resp) => {
                 if (err) {
                     return next(err)
                 } else {
@@ -202,7 +202,7 @@ let handlers = {
     results(req, res) {
         let jobId = req.params.jobId;
         if (req.body.status === 'FINISHED' || req.body.status === 'FAILED') {
-            getOutputs(jobId, (results) => {
+            agave.getOutputs(jobId, (results) => {
                 c.jobs.updateOne({jobId}, {$set: {agave: req.body, results}}, {}).then((err, result) => {
                     if (err) {res.send(err);}
                     else {res.send(result);}
@@ -228,8 +228,8 @@ let handlers = {
             if (status === 'FINISHED' || status === 'FAILED') {
                 res.send(job);
             } else {
-                agave.getJob(jobId, (err, resp) => {
-                    getOutputs(jobId, (results) => {
+                agave.api.getJob(jobId, (err, resp) => {
+                    agave.getOutputs(jobId, (results) => {
                         c.jobs.updateOne({jobId}, {$set: {agave: resp.body.result, results}}, {}, (err, result) => {
                             if (err) {res.send(err);}
                             else {res.send({agave: resp.body.result, results});}
@@ -337,7 +337,7 @@ let handlers = {
                         path = 'jobs/v2/' + jobId + '/outputs/media' + result.path;
                         let name = result.name;
 
-                        agave.getPath(path, (err, res, token) => {
+                        agave.api.getPath(path, (err, res, token) => {
                             let body = res.body;
                             if (!body || (body.status && body.status === 'error')) {
                                 // error from AGAVE
@@ -359,7 +359,7 @@ let handlers = {
             } else {
                 // download individual file
                 path = 'jobs/v2/' + jobId + '/outputs/media' + path;
-                agave.getPathProxy(path, res);
+                agave.api.getPathProxy(path, res);
             }
         });
 
@@ -411,103 +411,5 @@ let handlers = {
     }
 
 };
-
-// helper methods ----------------------------------------------------------
-
-function getOutputs (jobId, callback) {
-    let results = [];
-    agave.getJobOutput(jobId, (err, res) => {
-        let output = res.body.result;
-        if (output) {
-            // get main output files
-            for (let file of output) {
-                if ((file.name.indexOf('.err') > -1 || file.name.indexOf('.out') > -1) && file.length > 0) {
-                    results.push(file);
-                }
-            }
-        }
-        agave.getJobResults(jobId, (err, resp) => {
-            if (resp.body.result) {
-                results = results.concat(resp.body.result);
-            }
-            results = results.length > 0 ? results : null;
-            callback(results);
-        });
-    });
-}
-
-function submitJob (job, callback) {
-
-    let attempts = job.attempts ? job.attempts + 1: 1;
-
-    // form job body
-    let body = {
-        name:              'crn-automated-job',
-        appId:             job.appId,
-        batchQueue:        job.batchQueue,
-        executionSystem:   job.executionSystem,
-        maxRunTime:        '05:00:00',
-        memoryPerNode:     job.memoryPerNode,
-        nodeCount:         job.nodeCount,
-        processorsPerNode: job.processorsPerNode,
-        archive:           true,
-        archiveSystem:     'openfmri-archive',
-        archivePath:       null,
-        inputs: {},
-        parameters: job.parameters ? job.parameters : {},
-        notifications: [
-            {
-                url:        config.url + config.apiPrefix +  'jobs/${JOB_ID}/results',
-                event:      '*',
-                persistent: true
-            }
-        ]
-    };
-
-    // set input
-    let inputKey = job.input ? job.input : 'bidsFolder';
-    body.inputs[inputKey] = config.agave.storage + job.datasetHash;
-
-    // submit job
-    agave.createJob(body, (err, resp) => {
-
-        // handle submission errors
-        if (resp.body.status == 'error') {
-            let error = new Error(resp.body.message);
-            error.http_code = 400;
-            return callback(error, null);
-        }
-        if (resp.statusCode !== 200 && resp.statusCode !== 201) {
-            let error = new Error(resp.body ? resp.body : 'AGAVE was unable to process this job submission.');
-            error.http_code = resp.statusCode ? resp.statusCode : 503;
-            return callback(error, null);
-        }
-
-        // store job
-        c.jobs.insertOne({
-            jobId:             resp.body.result.id,
-            agave:             resp.body.result,
-
-            appId:             body.appId,
-            parameters:        body.parameters,
-            memoryPerNode:     body.memoryPerNode,
-            nodeCount:         body.nodeCount,
-            processorsPerNode: body.processorsPerNode,
-            batchQueue:        body.batchQueue,
-
-            appLabel:          job.appLabel,
-            appVersion:        job.appVersion,
-            datasetId:         job.datasetId,
-            datasetHash:       job.datasetHash,
-            userId:            job.userId,
-            parametersHash:    job.parametersHash,
-            snapshotId:        job.snapshotId,
-
-            attempts:          attempts
-        }, () => {
-            callback(null, resp.body);
-        });
-    });
-}
 
 export default handlers;
